Task: Genericize login form.
Add custom configuration setting to let
developers set an ADFS organization name
that will be referenced on the user/login.html.

I noticed that on some of the forks it was
common to change the login page to
remove references to NHS and switch to their
organization. Pulling this out into a config
setting felt like an easy solution to help
make this more usable for others without necessarily
needing to fork.