wip: Prevent new ADFS users from being auto created.
Let a configuration option determine how to handle
creating new ADFS users. For our organization we
want to controll who gains access at this point.
Adding a configuration boolean seems to be a
realatively unubtrusive way to do this.

It does seem to go against the default CKAN
setup that allows people to signup for basic access
and use but prevents them from adding/editing datasets
by default.

This may prove to be unnecessary in the future.