Task: Update _save_edit() action.
Prevent username and email from being changed
for ADFS users.

Validate user when trying to change password or
email (this really only applies to non-adfs users
as the ADFS users cannot update thier email and
do not recieve the password form fields) in the
template.

I did this to try and better align with CKAN's
default templates / logic. For this I added back
the current password field, and check to see if the
email has changed. I also display the user name and
email fields for ADFS users but disable them and
override any changes that may be submitted by the form
to prevent updates.

Some of the template updates were made in
[this commit](https://github.com/OpenGov-OpenData/ckanext-adfs/commit/bbf9e7f9b65af83e6283a55183f5344e9edf4020)
but I did not notice those before making my own changes.
They also cover a few other updates.