Improvement: Let user subscribe to email notifications.
Edit user form was not showing email notification
subscription checkbox option. This seems to be due
to aligning the edit_user_form.html to 2.8/master
instead of 2.7 which is currently being used.

Specifically the if block has changed between 2.7
and 2.8 that controls displaying the subscription
option.

While patching this I also made other changes to
small other items to better align CKAN versions
of the template to 2.7.