Bug: Deleted users can still login.
If a user is deleted and not purged (i.e. the state is deleted)
they can still login but got to a blank page. They are not able
to do anything as CKAN's authorization checks kick in but
the user should be rejected at the login form.

Found when testing to get the 403 error to appear.