Bug: Let adfs sysadmin edit other users.
Fix `email_changed` so that it wont be true for ADFS sysadmin
users trying to update another user. CKAN forces a sys admin
to give their password when make edits but with ADFS the passwords
aren't used. This allows the default behaviour for non-adfs users as well.

Fix name validation error for sysadmin adfs users trying to update another user.
For adfs users their name and email are blocked for editing. If someone does
try to edit these values they were replaced from session values. Changed this to
fetching the existing values from the DB as the session maybe a sysadmin.